libseccomp

+++ gck-rpc-dispatch.c
@@ -53,6 +53,14 @@
 #include <stdio.h>
 #include <syslog.h>
 
+#include <seccomp.h>
+//#include "seccomp-bpf.h"
+#ifdef DEBUG_SECCOMP
+# include "syscall-reporter.h"
+#endif /* DEBUG_SECCOMP */
+#include <fcntl.h> /* for seccomp init */
+#include <sys/mman.h>
+
 /* Where we dispatch the calls to */
 static CK_FUNCTION_LIST_PTR pkcs11_module = NULL;
 
@@ -97,6 +105,8 @@ static pthread_mutex_t pkcs11_dispatchers_mutex = PTHREAD_MUTEX_INITIALIZER;
 
 /* To be able to call C_Finalize from call_uninit. */
 static CK_RV rpc_C_Finalize(CallState *);
+
+static int _install_dispatch_syscall_filter(int use_tls);
 
 /* -----------------------------------------------------------------------------
  * LOGGING and DEBUGGING
@@ -2295,6 +2305,9 @@ static void *run_dispatch_thread(void *arg)
 	CallState *cs = arg;
 	assert(cs->sock != -1);
 
+	if (_install_dispatch_syscall_filter((cs->tls != NULL)))
+		return NULL;
+
 	run_dispatch_loop(cs);
 
 	/* The thread closes the socket and marks as done */
@@ -2407,7 +2420,7 @@ void gck_rpc_layer_inetd(CK_FUNCTION_LIST_PTR module)
  *
  * Returns -1 on failure, and the socket fd otherwise.
  */
-static int _get_listening_socket(char *host, char *port)
+static int _get_listening_socket(const char *proto, const char *host, const char *port)
 {
 	char hoststr[NI_MAXHOST], portstr[NI_MAXSERV];
 	struct addrinfo *ai, *first, hints;
@@ -2485,7 +2498,7 @@ static int _get_listening_socket(char *host, char *port)
 	}
 
 	snprintf(pkcs11_socket_path, sizeof(pkcs11_socket_path),
-		 (ai->ai_family == AF_INET6) ? "[%s]:%s" : "%s:%s", hoststr, portstr);
+		 (ai->ai_family == AF_INET6) ? "%s://[%s]:%s" : "%s://%s:%s", proto, hoststr, portstr);
 
  out:
 	freeaddrinfo(first);
@@ -2531,13 +2544,16 @@ int gck_rpc_layer_initialize(const char *prefix, CK_FUNCTION_LIST_PTR module)
 		 * TCP socket
 		 */
 		char *host, *port;
+		char proto[4]; /* strlen("tcp") and strlen("tls") */
+
+		snprintf(proto, sizeof(proto), "%s", prefix);
 
 		if (! gck_rpc_parse_host_port(prefix + 6, &host, &port)) {
 			free(host);
 			return -1;
 		}
 
-		if ((sock = _get_listening_socket(host, port)) == -1) {
+		if ((sock = _get_listening_socket(proto, host, port)) == -1) {
 			free(host);
 			return -1;
 		}
@@ -2598,7 +2614,9 @@ void gck_rpc_layer_uninitialize(void)
 	pkcs11_socket = -1;
 
 	/* Delete our unix socket */
-	if (pkcs11_socket_path[0])
+	if (pkcs11_socket_path[0] &&
+	    strncmp(pkcs11_socket_path, "tcp://", strlen("tcp://")) != 0 &&
+	    strncmp(pkcs11_socket_path, "tls://", strlen("tls://")) != 0)
 		unlink(pkcs11_socket_path);
 	pkcs11_socket_path[0] = 0;
 
@@ -2623,4 +2641,85 @@ void gck_rpc_layer_uninitialize(void)
 	pthread_mutex_unlock(&pkcs11_dispatchers_mutex);
 
 	pkcs11_module = NULL;
+}
+
+/*
+ * Reduce the syscalls allowed to a subset of the syscalls allowed for
+ * the parent thread.
+ */
+static int _install_dispatch_syscall_filter(int use_tls)
+{
+	int rc;
+
+#ifdef DEBUG_SECCOMP
+	rc = seccomp_init(SCMP_ACT_TRAP);
+#else
+	rc = seccomp_init(SCMP_ACT_KILL);
+#endif /* DEBUG_SECCOMP */
+	if (rc < 0)
+		goto failure_scmp;
+	/*
+	 * These are the basic syscalls needed to be able to use
+	 * the syscall-reporter to figure out the rest
+	 */
+      	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(write), 0);
+#ifdef DEBUG_SECCOMP
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(rt_sigreturn), 0);
+# ifdef __NR_sigreturn
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(sigreturn), 0);
+# endif
+#endif /* DEBUG_SECCOMP */
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(exit), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(exit_group), 0);
+
+	/*
+	 * Network related syscalls.
+	 */
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(read), 0);
+       	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(sendto), 0);
+
+	/*
+	 * TLS-PSK
+	 */
+	if (use_tls)
+		/* Allow open() of the TLS-PSK keyfile. */
+		seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(open), 1,
+				 SCMP_A1(SCMP_CMP_EQ, O_RDONLY | O_CLOEXEC));
+
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(close), 0);
+
+	/*
+	 * pthreads?
+	 */
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(madvise), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(mprotect), 1,
+			 SCMP_A2(SCMP_CMP_EQ, PROT_READ|PROT_WRITE));
+
+	/*
+	 * SoftHSM
+	 */
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(getcwd), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(stat), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(open), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(fcntl), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(fstat), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(lseek), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(access), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(fsync), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(unlink), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(ftruncate), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(select), 0);
+	seccomp_rule_add(SCMP_ACT_ALLOW, SCMP_SYS(futex), 0);
+
+	rc = seccomp_load();
+	if (rc < 0)
+		goto failure_scmp;
+	seccomp_release();
+
+	return 0;
+
+failure_scmp:
+	errno = -rc;
+	gck_rpc_warn("Seccomp filter initialization failed, errno = %u\n", errno);
+	return errno;
 }
